requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
python-dotenv
bcrypt # para encriptar contraseñas
streamlit_material # para darle un estilo a la app
graphviz



